backend/requirements.txt
Django==2.1.4
gunicorn
psycopg2
djangorestframework
django-cors-headers==2.4.0
celery
pandas
scipy
iapws
networkx